+++ crates/fuel-core/src/service/adapters/executor.rs
@@ -48,7 +48,7 @@ impl fuel_core_executor::ports::TransactionsSource for TransactionsSource {
     fn next(
         &self,
         gas_limit: u64,
-        transactions_limit: u16,
+        transactions_limit: u32,
         block_transaction_size_limit: u32,
     ) -> Vec<MaybeCheckedTransaction> {
         self.tx_pool
@@ -75,7 +75,7 @@ impl fuel_core_parallel_executor::ports::TransactionsSource for TransactionsSour
     fn get_executable_transactions(
         &mut self,
         gas_limit: u64,
-        tx_count_limit: u16,
+        tx_count_limit: u32,
         block_transaction_size_limit: u64,
         filter: Filter,
     ) -> TransactionSourceExecutableTransactions {
