+++ crates/services/executor/src/executor.rs
@@ -173,11 +173,11 @@ use alloc::{
 /// The maximum amount of transactions that can be included in a block,
 /// excluding the mint transaction.
 #[cfg(not(feature = "limited-tx-count"))]
-pub const fn max_tx_count() -> u16 {
-    u16::MAX.saturating_sub(1)
+pub const fn max_tx_count() -> u32 {
+    u32::MAX.saturating_sub(1)
 }
 #[cfg(feature = "limited-tx-count")]
-pub const fn max_tx_count() -> u16 {
+pub const fn max_tx_count() -> u32 {
     1024
 }
 
@@ -208,7 +208,7 @@ impl TransactionsSource for OnceTransactionsSource {
     fn next(
         &self,
         _: u64,
-        transactions_limit: u16,
+        transactions_limit: u32,
         _: u32,
     ) -> Vec<MaybeCheckedTransaction> {
         let mut lock = self.transactions.lock();
@@ -250,7 +250,7 @@ pub fn convert_tx_execution_result_to_preconfirmation(
     tx_id: TxId,
     tx_exec_result: &TransactionExecutionResult,
     block_height: BlockHeight,
-    tx_index: u16,
+    tx_index: u32,
 ) -> Preconfirmation {
     let tx_pointer = TxPointer::new(block_height, tx_index);
     let dynamic_outputs = tx
@@ -303,7 +303,7 @@ pub struct ExecutionData {
     pub coinbase: u64,
     pub used_gas: u64,
     pub used_size: u32,
-    pub tx_count: u16,
+    pub tx_count: u32,
     pub found_mint: bool,
     pub message_ids: Vec<MessageId>,
     pub tx_status: Vec<TransactionExecutionStatus>,
@@ -719,7 +719,7 @@ where
         mut self,
         transactions: Components<TxSource>,
         mut block_storage_tx: BlockStorageTransaction<D>,
-        start_idx: u16,
+        start_idx: u32,
         memory: &mut MemoryInstance,
     ) -> ExecutorResult<(Vec<Transaction>, ExecutionData)>
     where
